Expose auto gear module API globally
