+++ service-worker-assets.js
@@ -53,6 +53,7 @@
   "./legacy/scripts/modules/environment-bridge.js",
   "./legacy/scripts/modules/environment-context.js",
   "./legacy/scripts/modules/environment.js",
+  "./legacy/scripts/modules/features/auto-gear-rules.js",
   "./legacy/scripts/modules/features/backup.js",
   "./legacy/scripts/modules/features/feature-search.js",
   "./legacy/scripts/modules/features/help.js",
@@ -150,11 +151,13 @@
   "./src/scripts/modules/environment-bridge.js",
   "./src/scripts/modules/environment-context.js",
   "./src/scripts/modules/environment.js",
+  "./src/scripts/modules/features/auto-gear-rules.js",
   "./src/scripts/modules/features/backup.js",
   "./src/scripts/modules/features/feature-search.js",
   "./src/scripts/modules/features/help.js",
   "./src/scripts/modules/globals.js",
   "./src/scripts/modules/immutability.js",
+  "./src/scripts/modules/logging.js",
   "./src/scripts/modules/offline.js",
   "./src/scripts/modules/persistence.js",
   "./src/scripts/modules/registry.js",
